+++ backend/devices/bitbox02/handlers/handlers.go
@@ -55,7 +55,6 @@ type BitBox02 interface {
 	Product() bitbox02common.Product
 	GotoStartupSettings() error
 	RootFingerprint() ([]byte, error)
-	DeterministicEntropy() ([]byte, error)
 	BIP85AppBip39() error
 }